# doc/api/openapi.json
{
  "components": {
...
    "schemas": {
...
      "CreateRunRequest": {
        "properties": {
          "prompt": {
            "maxLength": 8000,
            "minLength": 3,
            "type": "string"
          },
          "providers": {
            "items": {
              "enum": [
                "openai",
                "anthropic",
                "google"
              ],
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          },
          "temperature": {
            "maximum": 2,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "prompt"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "code": {
            "enum": [
              "validation_failed",
              "unauthorized",
              "forbidden",
              "not_found",
              "conflict",
              "quota_exceeded",
              "rate_limited",
              "payment_required",
              "feature_unavailable",
              "budget_exhausted",
              "provider_unavailable",
              "not_configured",
              "internal_error"
            ],
            "type": "string"
          },
          "details": {
            "properties": {
              "billing": {
                "properties": {
                  "graceEndsAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "message": {
                    "type": "string"
                  },
                  "mode": {
                    "enum": [
                      "full",
                      "read-only"
                    ],
                    "type": "string"
                  },
                  "reason": {
                    "enum": [
                      "payment_past_due",
                      "subscription_canceled",
                      "subscription_paused",
                      null
                    ],
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "mode",
                  "reason",
                  "graceEndsAt",
                  "message"
                ],
                "type": "object"
              },
              "feature": {
                "type": "string"
              },
              "fields": {
                "items": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ],
                  "type": "object"
                },
                "type": "array"
              },
              "killSwitch": {
                "properties": {
                  "engaged": {
                    "type": "boolean"
                  },
                  "engagedAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "reason": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "releasedAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "scope": {
                    "enum": [
                      "global.spend"
                    ],
                    "type": "string"
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                },
                "required": [
                  "scope",
                  "engaged",
                  "reason",
                  "engagedAt",
                  "releasedAt",
                  "updatedAt"
                ],
                "type": "object"
              },
              "quota": {
                "properties": {
                  "ceiling": {
                    "minimum": 0,
                    "type": [
                      "integer",
                      "null"
                    ]
                  },
                  "limit": {
                    "enum": [
                      "monthly_runs",
                      "monthly_tokens",
                      "monthly_cost",
                      "concurrent_runs"
                    ],
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "plan": {
                    "enum": [
                      "free",
                      "pro",
                      "team",
                      "enterprise"
                    ],
                    "type": "string"
                  },
                  "remaining": {
                    "minimum": 0,
                    "type": [
                      "integer",
                      "null"
                    ]
                  },
                  "resetAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "upgradeTo": {
                    "enum": [
                      "free",
                      "pro",
                      "team",
                      "enterprise",
                      null
                    ],
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "used": {
                    "minimum": 0,
                    "type": "integer"
                  }
                },
                "required": [
                  "limit",
                  "used",
                  "ceiling",
                  "remaining",
                  "resetAt",
                  "plan",
                  "upgradeTo",
                  "message"
                ],
                "type": "object"
              },
              "rateLimit": {
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "limit": {
                    "exclusiveMinimum": 0,
                    "type": "integer"
                  },
                  "remaining": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "resetAt": {
                    "type": "string"
                  },
                  "retryAfterSeconds": {
                    "exclusiveMinimum": 0,
                    "type": "integer"
                  }
                },
                "required": [
                  "bucket",
                  "limit",
                  "remaining",
                  "resetAt",
                  "retryAfterSeconds"
                ],
                "type": "object"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "RunPage": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array"
          },
          "hasMore": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "data",
          "nextCursor",
          "hasMore"
        ],
        "type": "object"
      },
...
      "RunSummary": {
        "properties": {
          "canceledAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "candidateCount": {
            "type": "number"
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "confidence": {
            "type": [
              "number",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "createdByUserId": {
            "type": [
              "string",
              "null"
            ]
          },
          "deadlineAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasSynthesis": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "QUEUED",
              "FANNING_OUT",
              "SYNTHESIZING",
              "COMPLETE",
              "FAILED",
              "CANCELED"
            ],
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ]
          },
          "totalLatencyMs": {
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "createdByUserId",
          "prompt",
          "status",
          "error",
          "totalLatencyMs",
          "temperature",
          "createdAt",
          "completedAt",
          "deadlineAt",
          "canceledAt",
          "tags",
          "candidateCount",
          "hasSynthesis",
          "confidence"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "An API key: `Authorization: Bearer sce_live_…`. A browser session token from Clerk is also accepted, which is what the web app's playground uses.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "API support",
      "url": "https://github.com/gitdeepaks/self-consistency-answer-engine"
    },
    "description": "Ask several frontier models the same question, then have an evaluator merge their answers into one — with the agreements, the disagreements and a confidence score attached.\n\n**Authentication.** Every route takes `Authorization: Bearer sce_live_…`. Create a key in the web app under Settings → API keys. A key carries scopes; a request that needs one the key lacks is refused with 403 and `code: \"forbidden\"`.\n\n**Errors.** Every refusal is `{ code, message, details?, requestId }`. Branch on `code`, never on the HTTP status alone — 429 covers both a plan ceiling and a per-window request budget, and `details` tells them apart.\n\n**Retries.** Send an `Idempotency-Key` on every POST. A retry carrying the same key replays the original response instead of performing the write twice.\n\n**Runs are asynchronous.** `POST /v1/runs` returns as soon as the run is queued. Follow it with the SSE stream, or register a webhook and be told.",
    "license": {
      "name": "MIT"
    },
    "title": "Self-Consistency Answer Engine API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "description": "Newest first, cursor-paginated. Pass the previous page's `nextCursor` back as `cursor`; the value is opaque and must not be constructed by hand.",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "providers",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tags",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minConfidence",
            "required": false,
            "schema": {
              "maximum": 1,
              "minimum": 0,
              "type": [
                "number",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "A page of runs."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request did not match the expected shape."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No credential, or one that is expired or revoked."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential lacks the scope or role this route needs."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A plan ceiling or a per-window request budget was reached."
          }
        },
        "summary": "List runs",
        "tags": [
          "Runs"
        ]
      },
      "post": {
        "description": "Fans a prompt out across the model panel and returns immediately. The run is produced asynchronously; follow it with GET /v1/runs/{runId}/events or a run.completed webhook.\n\nSend an Idempotency-Key: a retry carrying the same key returns the original run with 200 rather than fanning out a second, identically expensive panel.",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "maxLength": 255,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "An earlier request with the same Idempotency-Key created this run."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "The run was created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request did not match the expected shape."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No credential, or one that is expired or revoked."
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The subscription cannot fund new work."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential lacks the scope or role this route needs."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A plan ceiling or a per-window request budget was reached."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The install-wide spend guard is engaged."
          }
        },
        "summary": "Start a run",
        "tags": [
          "Runs"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "This deployment",
      "url": "/v1"
    }
  ],
...
}